third_party/WebKit/LayoutTests/fast/forms/autofocus-keygen-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M54, around October 2016. See https://www.chromestatus.com/features/5716060992962560 for more details.

This tests whether the keygen element supports the autofocus attribute.

PASS document.activeElement is document.querySelector("keygen")
PASS document.activeElement.autofocus is true
PASS successfullyParsed is true

TEST COMPLETE
